+++ app/src/charts/TableChart.tsx
@@ -7,10 +7,23 @@ import TableContainer from "@material-ui/core/TableContainer";
 import TableHead from "@material-ui/core/TableHead";
 import TableRow from "@material-ui/core/TableRow";
 import { Row } from "../data/DatasetTypes";
-import { formatFieldValue } from "../utils/madlib/DisplayNames";
 import WarningRoundedIcon from "@material-ui/icons/WarningRounded";
 import { BreakdownCol, BREAKDOWN_VAR_DISPLAY_NAMES } from "../data/Breakdowns";
 import { MetricConfig } from "../data/MetricConfig";
+
+// Prints a formatted version of a field value based on the type specified by the field name
+export function formatFieldValue(nameOfField: string, value: any): string {
+  if (value === null || value === undefined) {
+    return "";
+  }
+  const formattedValue =
+    typeof value === "number" ? value.toLocaleString("en") : value;
+  const suffix =
+    nameOfField.endsWith("_pct") || nameOfField.endsWith("_pct_of_geo")
+      ? "%"
+      : "";
+  return `${formattedValue}${suffix}`;
+}
 
 function TableChart(props: {
   data: Row[];
